+++ R/geom-glycan.R
@@ -170,6 +170,7 @@ geom_glycan <- function(
     orient = orient,
     fuc_orient = style$fuc_orient,
     red_end = red_end,
+    red_end_length = style$red_end_length,
     edge_linewidth = style$edge_linewidth,
     node_linewidth = style$node_linewidth,
     node_size = style$node_size,
@@ -210,6 +211,7 @@ geom_glycan <- function(
 #' @param orient Drawing orientation passed to [glycanGrob()].
 #' @param fuc_orient Fuc-like triangle orientation passed to [glycanGrob()].
 #' @param red_end Reducing-end annotation passed to [glycanGrob()].
+#' @param red_end_length Reducing-end line length passed to [glycanGrob()].
 #' @param edge_linewidth Linkage linewidth passed to [glycanGrob()].
 #' @param node_linewidth Node-border linewidth passed to [glycanGrob()].
 #' @param node_size Node-size multiplier passed to [glycanGrob()].
@@ -228,6 +230,7 @@ geom_glycan <- function(
   orient = "left",
   fuc_orient = "flex",
   red_end = "",
+  red_end_length = 0.6,
   edge_linewidth = 0.8,
   node_linewidth = 0.8,
   node_size = 1,
@@ -276,7 +279,8 @@ geom_glycan <- function(
         node_linewidth = node_linewidth,
         node_size = node_size,
         font_family = font_family,
-        colors = colours
+        colors = colours,
+        red_end_length = red_end_length
       )
     )
   })